refactored type folder
