+++ src/js/engagement_view/src/components/engagementView/dynamicEngagementViewFeatures/utils/selectLens.tsx
@@ -6,7 +6,7 @@ import TableRow from "@material-ui/core/TableRow";
 
 import { useStyles } from '../styles';
 
-import { SelectLensProps } from "../types"
+import { SelectLensProps } from "../../../../types/DynamicEngagementViewTypes"
 
 export function SelectLens(props: SelectLensProps) {
     const classes = useStyles();
