+++ src/js/engagement_view/src/components/engagementView/dynamicEngagementViewFeatures/utils/toggleNodeTable.tsx
@@ -6,7 +6,7 @@ import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 
 import {
     ToggleNodeTableProps,
-} from "../types";
+} from "../../../../types/DynamicEngagementViewTypes";
 
 
 import {NodeDetails} from '../DynamicEngagementViewFeatures'
