+++ src/js/engagement_view/src/types/CustomTypes.tsx
@@ -189,9 +189,6 @@ export type ToggleLensTableState = {
     offset: number,
 }
 
-
-
-
 export type PaginationState = {
     first: number, 
     lenses: Lens[],
